refactor(tests): consolidate scattered test files (issue #1015)
Merge test_<module>/ directories into <module>/ directories, eliminating
duplicate test file structure across tests/vibe3/.

Changes:
- Move 14 non-conflicting test files from test_*/ to canonical directories
- Merge 3 conflicting file pairs:
  * test_issue_body_service.py: combine parse_projection_with_fallback tests
    with TestParseProjection/TestRenderProjection/TestMergeProjection classes
  * test_task_resume_operations.py: combine reset scenarios with clear/restore
    projection tests
  * test_supervisor.py: combine role module tests with profile resolution tests
- Move 4 root-level stray files to services/
- Clean up empty test_*/ directories
- Add __init__.py to adapters/ and config/ test directories

Verification:
- All 1089 tests pass (0 failures)
- Ruff lint checks pass
- No new mypy errors (pre-existing annotation warnings remain)
