+++ tests/vibe3/config/test_adapter_manifest.py
