+++ tests/vibe3/config/test_profile_convention.py
