+++ tests/vibe3/models/test_issue_body.py
