+++ tests/vibe3/services/test_bootstrap_context_service.py
