+++ tests/vibe3/services/test_convention_resolver.py
