+++ tests/vibe3/services/test_dispatch_error_propagation.py
