+++ tests/vibe3/services/test_flow_block_mixin.py
