+++ tests/vibe3/services/test_verdict_service.py
