+++ tests/vibe3/services/test_worktree_path_recording.py
